fix: don't push seed words in docs (it's against bdk philosophy)
