+++ docs/cookbook/keys-descriptors/seed-phrase.md
@@ -1,4 +1,4 @@
-# Wallet Recovery from Seeds
+# Recover Descriptors from Seeds
 
 BDK wallets require the use of descriptors, but recovery phrases (also called seed phrases) are a common and popular backup solution. Creating descriptors from a recovery phrase is a common workflow and BDK makes this easy with its _descriptor templates_, which are offered for common descriptors ([BIP 44/49/84/86](https://docs.rs/bdk_wallet/latest/bdk_wallet/descriptor/template/index.html)).
 
@@ -14,5 +14,5 @@ BDK wallets require the use of descriptors, but recovery phrases (also called se
 ### Example
 
 ```rust
---8<-- "examples/rust/wallet-recovery/src/main.rs"
+--8<-- "examples/rust/seed-phrase/src/main.rs"
 ```
