+++ examples/rust/seed-phrase/.gitignore
