+++ examples/rust/seed-phrase/Cargo.toml
@@ -5,4 +5,3 @@ edition = "2021"
 
 [dependencies]
 bdk_wallet = { version = "=1.0.0-beta.5", features = ["keys-bip39"] }
-bdk_esplora = { version = "=0.19.0", features = ["blocking"] }
